Update PS4DBG.Enums.cs
Renamed Enums.cs to PS4DBG.Enums.cs before moving it to the PS4DBG directory.
Added a comment to the VM_PROTECTIONS enum, and commented each one of it's members.
Added a comment to the WATCHPT_BREAKTYPE enum, and redid the pre-existing comments added to it's members.
Added a comment to the WATCHPT_LENGTH enum, and commented each one of it's members.